% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steuer_AfA.R
\name{AfA}
\alias{AfA}
\title{Abschreibung}
\usage{
AfA(
  ...,
  jahr = lubridate::year(lubridate::as_date(Sys.time())) - 1,
  konto_path = "C:/Users/wpete/Dropbox/2_Finanzen/Kontospiegel/2019/"
)
}
\arguments{
\item{...}{Liste mit Anlagen}

\item{jahr}{Jahr}

\item{konto_path}{Pfad}
}
\description{
linearen AfA
}
