% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stundenliste.R
\name{Anwesenheitszeiten}
\alias{Anwesenheitszeiten}
\title{Anwesenheitszeiten}
\usage{
Anwesenheitszeiten(
  monat = lubridate::month(Sys.Date()),
  jahr = lubridate::year(Sys.Date()),
  start = NULL,
  Taetigkeiten = list(list("1: Teamgruppenbesprechungen", 1),
    list("2: Student/innenbesprechstunden", 6), list("3: Fixe Lehrtermine", 0),
    list("4: Fixtermine an technischen Geräten", 0),
    list("5: Sonstige dienstliche Erfordernisse", 3)),
  Arbeitstage = c("Monday", "Tuesday", "Wednesday", "Thursday"),
  Woche = c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday"),
  Arbeitsfrei = setdiff(Woche, Arbeitstage),
  Wochenende = c("Saturday", "Sunday"),
  Urlaub = NULL
)
}
\arguments{
\item{monat}{Number}

\item{jahr}{Number}

\item{start}{Number}

\item{Taetigkeiten}{liste}

\item{Arbeitstage}{string}

\item{Woche}{string}

\item{Arbeitsfrei}{string}

\item{Wochenende}{string}

\item{Urlaub}{string}
}
\value{
data.frame
}
\description{
Anwesenheitszeiten
}
\examples{
\dontrun{
 Anwesenheitszeiten(monat = 12,
jahr = 2022,
start = 15,
Urlaub= c("2022-12-19",  "2022-12-30"))

  holidays()
  }
}
