% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prjc-arbeitszeit.R
\name{Arbz}
\alias{Arbz}
\title{Dummy fuer die Stundenliste}
\usage{
Arbz(task = NULL, time = NULL, ...)
}
\arguments{
\item{task, time}{Task als numeric und  Minuten}

\item{...}{nicht benutzt}
}
\value{
string
}
\description{
Dummy fuer die Stundenliste
}
\examples{
\dontrun{
Arbz()
}
}
