% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenplate-Angebot.R
\name{CreateAngebot}
\alias{CreateAngebot}
\title{Angebot erstellen}
\usage{
CreateAngebot(
  Name = "Vorname Name",
  Anrede = "Sehr geehrte Frau",
  Email = " ",
  Tel = " ",
  Adresse = " ",
  Datum = format(Sys.time(), "\%d.\%m.\%Y"),
  Folder = "C:/Users/wpete/Dropbox/1_Projekte",
  KNr = "000",
  x =
    "Position  Aufwand\\n  Einarbeiten_ins_Thema_und_Kommunikation  2\\n  Aufbereiten_der_Daten_und_Fehlerprüfung  1.5\\n  Deskriptive_Analyse_in_Form_von_Tabelle  0.5\\n  Deskriptive_Analyse_in_Form_von_Grafiken   3\\n  Inferenzstatistische_Analyse_(signifikanz_Test)  0.5\\n  Ergebnisbericht_mit_stichwortartigen_Kommentaren\\t 2\\n  ",
  stundensatz = 100,
  Betreff = "Aufwandseinschätzung",
  einleitung = "Vielen Dank für Ihre Anfrage!",
  text = "Ich unterbreite Ihnen hiermit folgendes Angebot: ",
  projekt = "Auswertung im Kontext einer wissenschaftlichen Arbeit",
  text2 = "Die nachfolgende Schätzung erfolgt auf Grund von Erfahrungswerten.",
  text3 =
    paste("Die Abrechnung erfolgt auf geleistete Arbeitsstunden, der Stundensatz beträgt ",
    stundensatz, "Euro.",
    "(Es wird keine Umsatzsteuer ausgewiesen, da die Umsätze  gemäß § 6 Abs. 1 Z 27 UStG unecht USt.-befreit sind.)"),
  closing =
    "Ich hoffe, dass Ihnen unser Angebot zusagt und  würde mich über Ihren Auftrag sehr freuen.",
  sep = "\\n\\n"
)
}
\arguments{
\item{Name}{character.}

\item{Anrede}{character.}

\item{Email}{character.}

\item{Tel}{character.}

\item{Adresse}{character.}

\item{Datum}{character. Automatisch}

\item{Folder}{character.}

\item{KNr}{character.}

\item{x}{character. Tabelle mit den Posten}

\item{stundensatz}{character.}

\item{Betreff}{character.}

\item{einleitung}{character.}

\item{text}{character.}

\item{projekt}{character.}

\item{text2}{character.}

\item{text3}{character.}

\item{closing}{character.}

\item{sep}{character.}
}
\value{
Folder Name
}
\description{
Angebot erstellen
}
