% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prjc-CreateProjekt.R, R/templat-Rechnung.R,
%   R/template-Stundenliste.R
\name{CreateProjekt}
\alias{CreateProjekt}
\alias{update_status_csv}
\alias{update_arbeitszeit}
\alias{insert_new_csv}
\alias{Rechnung}
\alias{rechnung_email}
\alias{Stundenliste}
\title{Projekt erstellen}
\usage{
CreateProjekt(
  KNr = NA,
  Name = "Romana Dampf",
  Email = "hd@gans.dampf.com",
  Telfon = "Tel 0000 12345678",
  Aufwand = "5-9 Stunden",
  Stundensatz = 95,
  Thema = "Forschung und Entwicklung",
  Kommentar = "Kein Kommentar",
  Anrede = "Sehr geehrte Frau",
  Betreff = "statistische Beratung im Zuge einer wissenschaftlichen Arbeit",
  Folder = "C:/Users/wpete/Dropbox/1_Projekte",
  Datum = format(Sys.time(), "\%d.\%m.\%Y"),
  Zeit = format(Sys.time(), "\%H:\%M"),
  Zwischenrechnung = 999,
  kunden_file = "C:/Users/wpete/Dropbox/1_Projekte/Verwaltung/Kunden.csv",
  bank = "TIROLER SPARKASSE",
  iban = IBAN(),
  bic = "SPIHAT22XXX"
)

update_status_csv(
  kunden_file = "C:/Users/wpete/Dropbox/1_Projekte/Verwaltung/Kunden.csv",
  KNr = 0L,
  Status = "unbekannt",
  Datum_Status = format(Sys.time(), "\%d.\%m.\%Y")
)

update_arbeitszeit(
  kunden_file = "C:/Users/wpete/Dropbox/1_Projekte/Verwaltung/Kunden.csv",
  KNr = 0L,
  Zeit = 0,
  Status = "in Arbeit",
  Datum_Status = format(Sys.time(), "\%d.\%m.\%Y")
)

insert_new_csv(
  kunden_file,
  KNr,
  Name,
  Email,
  Telfon,
  Aufwand,
  Stundensatz,
  Datum,
  Zeit,
  Status,
  Datum_Status = format(Sys.time(), "\%d.\%m.\%Y")
)

Rechnung(
  KNr = "000",
  Name = "Vorname Nachname",
  Telefon = "+43 123 456789",
  Email = "v.n@hotmail.com",
  Anrede = "Sehr geehrter Herr",
  Betreff = "statistische Beratung im Zuge einer wissenschaftlichen Arbeit",
  bank = "TIROLER SPARKASSE",
  iban = "AT89 2050 3033 XXXX XXXX",
  bic = "SPIHAXXXXXX",
  Adresse = "",
  stundenliste = "stundenliste.R",
  file_rechnung = NULL,
  ...
)

rechnung_email(
  KNr,
  Name,
  Email,
  Anrede,
  Euro = 0,
  Stundenliste,
  bank,
  iban,
  bic
)

Stundenliste(
  KNr,
  Name,
  Email,
  Telfon,
  Stundensatz,
  Anrede,
  Zwischenrechnung,
  bank,
  iban,
  bic,
  string_myswd,
  kunden_file
)
}
\arguments{
\item{KNr}{Kundennummer default ist NA}

\item{Name, Anrede, Email, Telfon, Adresse}{Kunde}

\item{Aufwand, Thema, Kommentar, Betreff, Stundensatz}{Projekt Daten}

\item{Folder}{Ort des Projektes}

\item{Datum, Zeit, Zwischenrechnung}{Allgemeine Parameter}

\item{kunden_file}{pfad}

\item{bank, iban, bic}{Bankverbindung}
}
\value{
nichts

character LaTex String
}
\description{
Erstellt ein neues Projekt mit der Ordnerstrucktur
und den R-Auswertungs-Files.
}
\examples{
stp25Project::Rechnung(
KNr = "123",
Name = "Hans Dampf",
Telefon = "+43 123 456789",
Email = "Hans_dampf@hotmail.com",
Anrede = "Sehr geehrter Herr",
bank = "TIROLER SPARKASSE",
iban = "AT89 2050 3033 XXXX XXXX",
bic = "SPIHAXXXXXX",
file_rechnung="Rechnung.qmd")

#  rmarkdown::render("Rechnung.qmd", encoding = "UTF-8")

}
