% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AZ.R
\name{extended_AZ}
\alias{extended_AZ}
\title{hier werden die Minuten in Zeit umgewandelt}
\usage{
extended_AZ(
  my_az,
  projektabschluss,
  order,
  zwischenrechnungen = FALSE,
  mynames = c("Datum", "Start", "Ende", "Projekt", "Task"),
  jetzt = Sys.time()
)
}
\description{
hier werden die Minuten in Zeit umgewandelt
}
