% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_projekt.R
\name{install_projekt}
\alias{install_projekt}
\title{Neues Projektordner erstellen}
\usage{
install_projekt(project = "000 Dummy",
  path = "C:/Users/wpete/Dropbox/1_Projekte", knr = 0,
  datum = date(), time = "08:08", name = "Name",
  email = "name@email.com", tel = "Tel", adr = "Adresse",
  h = "5-9 Stunden", thema = Thema, comment = "Test Dummy",
  euro = "76", anrede = "Sehr geehrte Damen und Herren",
  betreff = "statistische Beratung")
}
\arguments{
\item{project}{Name des projektes}

\item{path}{Pfad default Dropbox/1_Projekte}

\item{knr, time, name, tel, email, adr, h, thema, euro, anrede, betreff}{Kunden-Daten}

\item{datum}{Datum}

\item{comment}{Beschreibung}
}
\description{
Neues Projektordner erstellen
}
