% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rechnung.R
\name{rechnung_email}
\alias{rechnung_email}
\title{Rechnung erstellen}
\usage{
rechnung_email(Name = "Hans Dampf", Email = "h.dampf@gmail.com",
  Anrede = "Hallo", Euro = 12, RNr = "0815", Stundenliste = "")
}
\arguments{
\item{Name, Anrede, Email, Euro, RNr, Stundenliste}{Kunde}
}
\value{
text
}
\description{
Rechnung erstellen
}
