% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{small_project}
\alias{small_project}
\title{title: ", project,"
author: Wolfgang Peter
output:
  pdf_document
keep_tex: yes
lang: de-DE
---}
\usage{
small_project(project = "dummy", datum = "Fri May 10 08:46:24 2019",
  myswd = NA, Rdata = "DF.Rdata")
}
\description{
title: ", project,"
author: Wolfgang Peter
output:
  pdf_document
keep_tex: yes
lang: de-DE
---
}
