% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steuer.R
\name{Steuer}
\alias{Steuer}
\title{Einkommensteuer}
\usage{
Steuer(
  jahr = lubridate::year(lubridate::as_date(Sys.time())) - 1,
  konto = "konto-spiegel.xlsx",
  AfA = data.frame(Pos = c("Anlagen", "Instandsetzung"), Kennzahl = c(9130, 9150), Betrag
    = c(0, 0)),
  konto_path = "C:/Users/wpete/Dropbox/2_Finanzen/Kontospiegel/2023/",
  select = NULL,
  Versicherung.SVA = NULL,
  Reisespesen = c(Bus = 0, Taxi = 0),
  Betriebsausgaben = NULL,
  Versicherungspraemien = 0,
  Steuerberater = 0,
  pauschale = 0.06,
  privat = c("AMAZON", "Delinat", "GOOGLE", "Deutscher Pressevertrieb", "TIWAG",
    "Privatentnahme", "Krabe", "GIRSTMAIR", "VAILLANT", "Wolfgang", "SCHULER", "OEBB",
    "AUDIBLE", "AFA"),
  pattern_debit = "Debit",
  sva = "Sozialvers.Anstalt",
  versicherung = "UNIQA",
  kirche = "Bischöfliche",
  spenden = c("ÖRK", "Kinderpatenschaft")
)
}
\arguments{
\item{jahr}{Jahr}

\item{konto}{character, xls-Datei direkt aus der Bank exportiere Kontospiegel}

\item{AfA}{data.frame. mit der Funktion AfA() erzeugte Daten.}

\item{konto_path}{Pfad zum Kontospiegel}

\item{select}{Postitionen für c(Buchungsdatum = 1, Partnername =2, Betrag = 11)}

\item{Versicherung.SVA}{numeric. Zahlungsbestätigung der SAV}

\item{Reisespesen, Betriebsausgaben, Versicherungspraemien, Steuerberater}{numeric.}

\item{pauschale}{default 0.06  6 Prozent Pauschalierung fuer Kleinunternehmer}

\item{pattern_debit, privat, versicherung, kirche, spenden, sva}{character. pattern}
}
\value{
data.frame und speichern der Berechnungen als csv-daten
}
\description{
https://finanzonline.bmf.gv.at/fon/login.do
+4369981530117
}
\details{
grundsaetzlich bis 30. April des Folgejahres bis 30. Juni bei uebermittlung 
auf elektronischem Wege


Branche: 73.2 Markt- und Meinungsforschung
Gewinnermittlungsarten
Vollständige Einnahmen-Ausgaben-Rechnung gemäß § 4 Abs. 3 => Ja
USt-Bruttosystem => Ja

Angaben zur Kommunalsteuer

Gemeindekennziffer: 70364 
Bemessungsgrundlage: 0
Gesamtbetrag der Bemessungsgrundlage: 0
Gesamtbetrag Kommunalsteuer: 0

Taetigkeit als  Liebhaberei 
Wenn die Vermietung oder Verpachtung ueber einen laengeren Zeitraum keinen Gewinn
erwarten laesst, sind diese Einkuenfte steuerlich unbeachtlich. 
Das heißt, dass Verluste aus
dieser Taetigkeit mit anderen positiven Einkuenften nicht ausgeglichen werden duerfen.
Sollte sich ausnahmsweise ein Gewinn ergeben, ist dieser nicht steuerpflichtig.
Zu unterscheiden ist dabei zwischen der so genannten  kleinen  Vermietung wie z.B. der
Vermietung von Ein- und Zweifamilienhaeusern, Eigentumswohnungen oder Bungalows und
der  großen  Vermietung. Das ist die Vermietung von mindestens drei Wohneinheiten,
betrifft also vor allem Miethaeuser (siehe dazu das Infoblatt  Liebhaberei imSteuerrecht ).
 


Steuer(konto = "AT892050303302884626_2025-01-01_2025-12-31.xlsx",
      AfA = afa,
      konto_path = "C:/Users/wpete/luckycloud/2_Finanzen/Kontospiegel/2025/",
      #
      Versicherung.SVA = 2483.89,
      sva = c("Sozialvers.Anstalt"),
      #
      select=  c(Buchungsdatum = 1, Partnername =2, Betrag = 11)
 )
}
\examples{
 \dontrun{
afa<- 
AfA(
  list(
    nr = 1, datum = "26.01.2012",
    pos = "PC L412 Windert GmbH, Bad Oldeslose",
    type = "A",
    euro = 463.95, nd = 3, rs = 1),
  list(
    nr = 33, datum = "09.11.2012",
    pos = "Buero Fenster+Tuer EA-Ceramic e.U. Hallerstraße 35, Ibk",
    type = "I",
    euro = 3517.20, nd = 10, rs = 0.5),
  list(
    nr = 34, datum = "03.12.2012",
    pos = "Buero Fussboden-Sanierung EA-Ceramic e.U. Hallerstraße 35. Ibk",
    type = "I",
    euro = 2271.71, nd = 10, rs = 0.5),
  list(
    nr = 35, datum = "03.12.2012",
    pos = "Buero Fussboden-Heizung EA-Ceramic e.U. Hallerstraße 35. Ibk",
    type = "I",
    euro = 912.48, nd = 10, rs = 0.5),
  list(
    nr = 21694, datum = "23.01.2018",
    pos = "Thinkpad T450s Matthias Rimkus",
    type = "A",
    euro = 475, nd = 3, rs = 1),
  list(
    nr = 130603,  datum = "07.06.2013",
    pos = "Bueromöbel Leitgeb&leitgeb 6020 Innsbruck",
    type = "I",
    euro = 8436.00, nd = 10, rs = 0.5),
  jahr=2019,
  konto_path = "C:/Users/wpete/Dropbox/2_Finanzen/Kontospiegel/2019/" 
)

Steuer(konto = "AT892050303302884626_2025-01-01_2025-12-31.xlsx",
      AfA = afa,
      konto_path = "C:/Users/wpete/luckycloud/2_Finanzen/Kontospiegel/2025/",
      #Zahlungsbestätigung der SAV
      Versicherung.SVA = 2483.89,
      #direkt aus der Bank exportiere Kontospiegel
      select=  c(Buchungsdatum = 1, Partnername =2, Betrag = 11)
 )
        
}

}
