% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vertrag.R
\name{Vertrag}
\alias{Vertrag}
\title{Vertrag Templat tEMPLAT}
\usage{
Vertrag(
  name = "Vorname Nachname",
  adr = "bitte ergänzen",
  tel = "bitte ergänzen",
  email = "bitte ergänzen",
  knr = "0",
  euro = "76",
  h = " 5-9 Stunden",
  Betreff = "statistische Beratung",
  zwischenrechnung = 400,
  BANK = "TIROLER SPARKASSE",
  IBAN = IBAN(),
  BIC = "SPIHAT22XXX"
)
}
\arguments{
\item{name, adr, tel, email, knr, euro, h}{Kunde}
}
\value{
string
}
\description{
Vertrag Templat tEMPLAT
}
