% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AZ.R
\name{AZ}
\alias{AZ}
\title{Arbeitszeit}
\usage{
AZ(Lines, sep = "\\\\t", projektabschluss = 7, ...)
}
\arguments{
\item{Lines}{Arbeitszeiten als Text}

\item{sep}{Tabulator}

\item{projektabschluss}{Zeit für Projektabschluss}

\item{...}{nicht Benutzt}
}
\value{
data.frame
}
\description{
Arbeitszeit
}
