% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prjc-CreateProjekt.R, R/prjc-install_projekt.R,
%   R/templat-Rechnung.R, R/template-Stundenliste.R, R/template-Vertrag.R
\name{CreateProjekt}
\alias{CreateProjekt}
\alias{update_status_csv}
\alias{update_arbeitszeit}
\alias{insert_new_csv}
\alias{install_projekt}
\alias{Rechnung}
\alias{rechnung_email}
\alias{Stundenliste}
\alias{Vertrag}
\title{Projekt erstellen}
\usage{
CreateProjekt(
  KNr = NA,
  Name = "Romana Dampf",
  Email = "hd@gans.dampf.com",
  Telfon = "Tel 0000 12345678",
  Aufwand = "5-9 Stunden",
  Stundensatz = 95,
  Thema = "Forschung und Entwicklung",
  Kommentar = "Kein Kommentar",
  Anrede = "Sehr geehrte Frau",
  Betreff = "statistische Beratung im Zuge einer wissenschaftlichen Arbeit",
  Folder = "C:/Users/wpete/Dropbox/1_Projekte",
  Datum = format(Sys.time(), "\%d.\%m.\%Y"),
  Zeit = format(Sys.time(), "\%H:\%M"),
  Zwischenrechnung = 999,
  Status = "AGB",
  kunden_file = "C:/Users/wpete/Dropbox/1_Projekte/Verwaltung/Kunden.csv",
  bank = "TIROLER SPARKASSE",
  iban = IBAN(),
  bic = "SPIHAT22XXX"
)

update_status_csv(
  kunden_file = "C:/Users/wpete/Dropbox/1_Projekte/Verwaltung/Kunden.csv",
  KNr = 0L,
  Status = "unbekannt",
  Datum_Status = format(Sys.time(), "\%d.\%m.\%Y")
)

update_arbeitszeit(
  kunden_file = "C:/Users/wpete/Dropbox/1_Projekte/Verwaltung/Kunden.csv",
  KNr = 0L,
  Zeit = 0,
  Status = "in Arbeit",
  Datum_Status = format(Sys.time(), "\%d.\%m.\%Y")
)

insert_new_csv(
  kunden_file,
  KNr,
  Name,
  Email,
  Telfon,
  Aufwand,
  Stundensatz,
  Datum,
  Zeit,
  Status,
  Datum_Status = format(Sys.time(), "\%d.\%m.\%Y")
)

install_projekt(
  Projekt,
  KNr,
  Name,
  Email,
  Telfon,
  Aufwand,
  Stundensatz,
  Thema,
  Kommentar,
  Anrede,
  Betreff,
  Folder,
  Datum,
  Zeit,
  Zwischenrechnung,
  Status,
  kunden_file = "C:/Users/wpete/Dropbox/1_Projekte/Verwaltung/Kunden.csv",
  bank,
  iban,
  bic,
  ...
)

Rechnung(KNr, Name, Email, Telfon, Anrede, Betreff, bank, iban, bic)

rechnung_email(
  KNr,
  Name,
  Email,
  Anrede,
  Euro = 0,
  Stundenliste,
  bank,
  iban,
  bic
)

Stundenliste(
  KNr,
  Name,
  Email,
  Telfon,
  Stundensatz,
  Anrede,
  Zwischenrechnung,
  bank,
  iban,
  bic,
  string_myswd
)

Vertrag(
  KNr,
  Name,
  Email,
  Telfon,
  Aufwand,
  Stundensatz,
  Anrede,
  Betreff,
  Zwischenrechnung,
  bank,
  iban,
  bic
)
}
\arguments{
\item{KNr, save_KNr}{Kundennummer normal ist NA}

\item{Name, Anrede, Email, Telfon, Adresse}{Kunde}

\item{Aufwand, Thema, Kommentar, Betreff, Stundensatz}{Projekt Daten}

\item{Datum, Zeit, Folder}{Allgemeine Parameter}

\item{kunden_file}{pfad}
}
\value{
nichts
}
\description{
Erstellt ein neues Projekt mit der Ordnerstrucktur
und den R-Auswertungs-Files.
}
