% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateProjekt.R
\name{CreateProjekt}
\alias{CreateProjekt}
\title{Projekt erstellen}
\usage{
CreateProjekt(
  Name = "Romana Dampf",
  Anrede = "Sehr geehrte Frau",
  Email = "hd@gans.dampf.com",
  Tel = "Tel 0000 12345678",
  Adresse = "Strasse, Ort",
  Aufwand = "5-9 Stunden",
  Thema = "Forschung und Entwicklung",
  Kommentar = "Kein Kommentar",
  Betreff = "statistische Beratung im Zuge einer wissenschaftlichen Arbeit",
  Stundensatz = 76,
  Datum = format(Sys.time(), "\%d.\%m.\%Y"),
  Zeit = format(Sys.time(), "\%H:\%M"),
  Folder = "C:/Users/wpete/Dropbox/1_Projekte",
  KNr = NA,
  save_KNr = is.na(KNr),
  kunden_file = "C:/Users/wpete/Dropbox/1_Projekte/Verwaltung/Kunden.csv",
  FunktionsTest = FALSE,
  Templat = "small",
  zwischenrechnung = 400,
  BANK = "TIROLER SPARKASSE",
  IBAN = IBAN(),
  BIC = "SPIHAT22XXX"
)
}
\arguments{
\item{Name, Anrede, Email, Tel, Adresse}{Kunde}

\item{Aufwand, Thema, Kommentar, Betreff, Stundensatz}{Projekt Daten}

\item{Datum, Zeit, Folder}{Allgemeine Parameter}

\item{KNr, save_KNr}{Kundennummer normal ist NA}

\item{kunden_file, FunktionsTest}{Datenbank}
}
\value{
nichts
}
\description{
Erstellt ein neues Projekt mit der Ordnerstrucktur
und den R-Auswertungs-Files.
}
