% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateProjekt.R
\name{CreateProjekt}
\alias{CreateProjekt}
\title{Neues Projekt erstellen}
\usage{
CreateProjekt(project = "000 Dummy",
  path = "C:/Users/wpete/Dropbox/1_Projekte", knr = 0,
  datum = date(), time = "08:08", name = "Name",
  email = "name@email.com", tel = "Tel", adr = "Adresse",
  h = "5-9 Stunden", thema = Thema, comment = "Test Dummy",
  euro = "76", anrede = "Sehr geehrte Damen und Herren",
  Betreff = "statistische Beratung")
}
\arguments{
\item{project}{Name des projektes}

\item{path}{Pfad default Dropbox/1_Projekte}

\item{knr, time, name, tel, email, adr, h, thema, euro, anrede, Betreff}{Kunden-Daten}

\item{datum}{Datum}

\item{comment}{Beschreibung}
}
\description{
Erstellt ein neues Projekt mit der Ordnerstrucktur
 und den R-Auswertungs-Files.
}
\examples{
## Not run:
}
