% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-GetData.R
\name{Get_Data}
\alias{Get_Data}
\title{title: Daten Aufbereiten
author: Wolfgang Peter
output:
  html_document
---}
\usage{
Get_Data(
  myswd = " setwd('C:/Users/wpete/Dropbox/1_Projekte/001 Dummy')",
  Rdata = "DF.Rdata"
)
}
\description{
title: Daten Aufbereiten
author: Wolfgang Peter
output:
  html_document
---
}
