% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rechnung.R
\name{Rechnung}
\alias{Rechnung}
\title{Rechnungs-Templat}
\usage{
Rechnung(
  KNr = 0,
  Name = "Vorname Name",
  Email = " ",
  Tel = " ",
  Adresse = " ",
  Anrede = "Sehr geehrte Frau",
  Betreff = "Beratung",
  BANK = "TIROLER SPARKASSE",
  IBAN = IBAN(),
  BIC = "SPIHAT22XXX"
)
}
\arguments{
\item{KNr, Name, Email, Tel, Adresse, Anrede}{KundenDaten}
}
\value{
string
}
\description{
Rechnungs-Templat
}
