% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{small_project}
\alias{small_project}
\title{title: ", project,"
author: Wolfgang Peter
output:
  html_document
keep_tex: yes
lang: de-DE
---}
\usage{
small_project(
  project = "dummy",
  datum = "",
  myswd = " setwd('C:/Users/wpete/Dropbox/1_Projekte/001 Dummy')",
  Rdata = "DF.Rdata"
)
}
\description{
title: ", project,"
author: Wolfgang Peter
output:
  html_document
keep_tex: yes
lang: de-DE
---
}
