% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steuer.R
\name{Steuer}
\alias{Steuer}
\title{Einkommensteuer}
\usage{
Steuer(
  jahr = lubridate::year(lubridate::as_date(Sys.time())) - 1,
  konto = "umsaetze2019.csv",
  SVA = NA,
  Betriebsausgaben = c(VISA = 0, Buecher = 0),
  Reisespesen = c(Bus = 0, Taxi = 0),
  Versicherungspraemien = 0,
  Steuerberater = 0,
  pauschale = 0.06,
  AfA = data.frame(Pos = c("Anlagen", "Instandsetzung"), Kennzahl = c(9130, 9150), Betrag
    = c(0, 0)),
  konto_path = "C:/Users/wpete/Dropbox/2_Finanzen/Kontospiegel/2019/",
  eigene_buchung_konto = c("AT573626000100540567", "AT573626000000540567"),
  visa = "VISA-RECHNUNG",
  versicherung = c("UNIQA"),
  privat = c("Delinat", "Lebensmitteltechnik", "Deutscher Pressevertrieb", "TIWAG",
    "Privatentnahme", "Hechenblaikner", "0006000079753746", "Wolfgang Peter"),
  kirche = "Bischöfliche",
  spenden = c("ÖRK", "Kinderpatenschaft"),
  sva = "AT69ZZZ00000003350"
)
}
\arguments{
\item{jahr}{Jahr}

\item{konto}{Kontospiegel}

\item{SVA, Betriebsausgaben, Reisespesen, Versicherungspraemien, Steuerberater}{Summe,}

\item{pauschale}{Pauschalierung für Kleinunternehmer}

\item{AfA, afa_anlagen, afa_instandsetzung}{Abschreibung}

\item{konto_path}{Pfad}

\item{eigene_buchung_konto, visa, versicherung, privat, kirche, spenden, sva}{Such-String fuer die jeweilige Buchung}
}
\value{
data.frame
}
\description{
https://finanzonline.bmf.gv.at/fon/login.do
+4369981530117
}
\details{
grundsätzlich bis 30. April des Folgejahres bis 30. Juni bei Übermittlung auf elektronischem Wege

 

Önace Code	Text	Tätigkeit 73.	Markt- und Meinungsforschung


Angaben zur Kommunalsteuer

Gemeindekennziffer: 70364 
Bemessungsgrundlage: 0
Gesamtbetrag der Bemessungsgrundlage: 0
Gesamtbetrag Kommunalsteuer: 0

Tätigkeit als  Liebhaberei 
Wenn die Vermietung oder Verpachtung über einen längeren Zeitraum keinen Gewinn
erwarten lässt, sind diese Einkünfte steuerlich unbeachtlich. Das heißt, dass Verluste aus
dieser Tätigkeit mit anderen positiven Einkünften nicht ausgeglichen werden dürfen.
Sollte sich ausnahmsweise ein Gewinn ergeben, ist dieser nicht steuerpflichtig.
Zu unterscheiden ist dabei zwischen der so genannten  kleinen  Vermietung wie z.B. der
Vermietung von Ein- und Zweifamilienhäusern, Eigentumswohnungen oder Bungalows und
der  großen  Vermietung. Das ist die Vermietung von mindestens drei Wohneinheiten,
betrifft also vor allem Miethäuser (siehe dazu das Infoblatt  Liebhaberei im
                                    Steuerrecht ).
}
\examples{
 \dontrun{
afa<- 
AfA(
  list(
    nr = 1, datum = "26.01.2012",
    pos = "PC L412 Windert GmbH, Bad Oldeslose",
    type = "A",
    euro = 463.95, nd = 3, rs = 1),
  list(
    nr = 33, datum = "09.11.2012",
    pos = "Büro Fenster+Tür EA-Ceramic e.U. Hallerstraße 35, Ibk",
    type = "I",
    euro = 3517.20, nd = 10, rs = 0.5),
  list(
    nr = 34, datum = "03.12.2012",
    pos = "Büro Fussboden-Sanierung EA-Ceramic e.U. Hallerstraße 35. Ibk",
    type = "I",
    euro = 2271.71, nd = 10, rs = 0.5),
  list(
    nr = 35, datum = "03.12.2012",
    pos = "Büro Fussboden-Heizung EA-Ceramic e.U. Hallerstraße 35. Ibk",
    type = "I",
    euro = 912.48, nd = 10, rs = 0.5),
  list(
    nr = 21694, datum = "23.01.2018",
    pos = "Thinkpad T450s Matthias Rimkus",
    type = "A",
    euro = 475, nd = 3, rs = 1),
  list(
    nr = 130603,  datum = "07.06.2013",
    pos = "Büromöbel Leitgeb&leitgeb 6020 Innsbruck",
    type = "I",
    euro = 8436.00, nd = 10, rs = 0.5),
  jahr=2019,
  konto_path = "C:/Users/wpete/Dropbox/2_Finanzen/Kontospiegel/2019/" 
)

Steuer(
  konto="umsaetze2019.csv", 
  AfA = afa,
  Versicherungspraemien = c(Zuericher = 460.56, Uniqa = 0),
  jahr=2019,
  konto_path = "C:/Users/wpete/Dropbox/2_Finanzen/Kontospiegel/2019/" 
)
}
}
