% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steuer_neu.R
\name{Steuer}
\alias{Steuer}
\title{Einkommensteuer}
\usage{
Steuer(
  jahr = lubridate::year(lubridate::as_date(Sys.time())) - 1,
  konto = "konto-spiegel.xlsx",
  konto_path = "C:/Users/wpete/Dropbox/2_Finanzen/Kontospiegel/2023/",
  Versicherung.SVA = NULL,
  privat = c("AMAZON", "Delinat", "GOOGLE", "Deutscher Pressevertrieb", "TIWAG",
    "Privatentnahme", "Krabe", "GIRSTMAIR", "VAILLANT", "Wolfgang", "SCHULER"),
  sva = c("Sozialversicherungsanstalt"),
  pattern_debit = "Debit",
  Reisespesen = c(Bus = 0, Taxi = 0),
  Betriebsausgaben = NULL,
  Versicherungspraemien = 0,
  Steuerberater = 0,
  pauschale = 0.06,
  AfA = data.frame(Pos = c("Anlagen", "Instandsetzung"), Kennzahl = c(9130, 9150), Betrag
    = c(0, 0)),
  versicherung = c("UNIQA"),
  kirche = "Bischöfliche",
  spenden = c("ÖRK", "Kinderpatenschaft")
)
}
\arguments{
\item{jahr}{Jahr}

\item{konto}{Kontospiegel}

\item{konto_path}{Pfad}

\item{Betriebsausgaben}{data.frame
   Betriebsausgaben = data.frame(
datum = "2022-11-07",
txt = "Amazon OKI B432dn A4-Schwarzweißdrucker (Duplex, Netzwerk)",
euro = -265.40,
id = 999,
einnahmen = NA,
ausgaben  = -265.40,
Pos = "Betriebsausgaben"
)}

\item{pauschale}{Pauschalierung fuer Kleinunternehmer}

\item{AfA, afa_anlagen, afa_instandsetzung}{Abschreibung}

\item{SVA, Reisespesen, Versicherungspraemien, Steuerberater}{Summe,}

\item{eigene_buchung_konto, visa, versicherung, privat, kirche, spenden, sva}{Such-String fuer die jeweilige Buchung}
}
\value{
data.frame
}
\description{
https://finanzonline.bmf.gv.at/fon/login.do
+4369981530117
}
\details{
grundsaetzlich bis 30. April des Folgejahres bis 30. Juni bei uebermittlung auf elektronischem Wege

 

Önace Code	Text	Taetigkeit 73.	Markt- und Meinungsforschung


Angaben zur Kommunalsteuer

Gemeindekennziffer: 70364 
Bemessungsgrundlage: 0
Gesamtbetrag der Bemessungsgrundlage: 0
Gesamtbetrag Kommunalsteuer: 0

Taetigkeit als  Liebhaberei 
Wenn die Vermietung oder Verpachtung ueber einen laengeren Zeitraum keinen Gewinn
erwarten laesst, sind diese Einkuenfte steuerlich unbeachtlich. Das heißt, dass Verluste aus
dieser Taetigkeit mit anderen positiven Einkuenften nicht ausgeglichen werden duerfen.
Sollte sich ausnahmsweise ein Gewinn ergeben, ist dieser nicht steuerpflichtig.
Zu unterscheiden ist dabei zwischen der so genannten  kleinen  Vermietung wie z.B. der
Vermietung von Ein- und Zweifamilienhaeusern, Eigentumswohnungen oder Bungalows und
der  großen  Vermietung. Das ist die Vermietung von mindestens drei Wohneinheiten,
betrifft also vor allem Miethaeuser (siehe dazu das Infoblatt  Liebhaberei im
                                    Steuerrecht ).
}
