% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stundenliste.R
\name{Stundenliste}
\alias{Stundenliste}
\title{Stundenliste}
\usage{
Stundenliste(
  euro = 85,
  myswd = "",
  knr,
  name,
  email,
  tel,
  adr,
  anrede,
  betreff,
  zwischenrechnung = 400,
  BANK = "BANK",
  IBAN = "IBAN",
  BIC = "BIC"
)
}
\arguments{
\item{euro}{Zahl}
}
\value{
text
title: Stundenliste
author: Wolfgang Peter
output:
  pdf_document: default
---
}
\description{
Stundenliste
}
