% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stundenliste.R
\name{Stundenliste}
\alias{Stundenliste}
\title{Stundenliste}
\usage{
Stundenliste(euro = 76, myswd = "", knr, name, email, tel, adr, anrede,
  betreff)
}
\arguments{
\item{euro}{Zahl}
}
\value{
text
title: Stundenliste
author: Wolfgang Peter
output:
  pdf_document: default
---
}
\description{
Stundenliste
}
